Significant modifications to the MutationObserver monitoring tiddler changes, which should provide improved performance while also avoiding issues like bug #1 (re. the duplication of draft titles with each MathJax typesetting call).
At present, there are two major bugs with this new version of the script that prevent it from functioning. First and foremost, the new, pared down MutationObserver, riverObserver, is not reacting to edits as anticipated. This is almost certainly due to the fact that it is now only observing characterData mutations (although I'm still familiarizing myself with these new-fangled MutationObservers, and it's unclear to me why a change to the value of a textarea wouldn't constitute a characterData mutation in the river's subtree), but the new callback function will require some massaging to deal with other types of mutations.

The other issue is one of scope, where it seems the callback function doesn't see the getAncestorTiddler and tiddlerIsADraft functions defined above in the file. This doesn't seem in the spirit of lexicographic scoping to me, so I'm assuming there's an easy fix here.

This commit should in no way shape or form be considered a patch for the bugs in this branch; I'm mostly committing progress before a potential rewrite of the (currently non-functional) mutation observer and its callback.